# obj/Debug/net10.0/EndpointInfo/ClyvoDayApiDocker.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Projeto CLYVO DAY - Documentação OpenAPI",
    "description": "Documentação da API CLYVO DAY - Plataforma de continuidade de cuidado veterinário.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:7037"
    },
    {
      "url": "http://localhost:5278"
    }
  ],
  "paths": {
    "/api/AnimalEstimacao": {
      "get": {
        "tags": [
          "AnimalEstimacao"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "AnimalEstimacao"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnimalEstimacao"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AnimalEstimacao"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AnimalEstimacao"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/AnimalEstimacao/{id}": {
      "get": {
        "tags": [
          "AnimalEstimacao"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "AnimalEstimacao"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/AnimalEstimacao/{id}/atualizapeso": {
      "put": {
        "tags": [
          "AnimalEstimacao"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
                "type": [
                  "number",
                  "string"
                ],
                "format": "double"
              }
            },
            "text/json": {
              "schema": {
                "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
                "type": [
                  "number",
                  "string"
                ],
                "format": "double"
              }
            },
            "application/*+json": {
              "schema": {
                "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
                "type": [
                  "number",
                  "string"
                ],
                "format": "double"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Clinica": {
      "get": {
        "tags": [
          "Clinica"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Clinica"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Clinica"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Clinica"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Clinica"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Clinica"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Clinica"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Clinica"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Clinica/{id}": {
      "get": {
        "tags": [
          "Clinica"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Clinica"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clinica"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clinica"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Clinica"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/EventoCuidado": {
      "get": {
        "tags": [
          "EventoCuidado"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventoCuidado"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventoCuidado"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventoCuidado"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "EventoCuidado"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventoCuidado"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventoCuidado"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventoCuidado"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/EventoCuidado/{id}": {
      "get": {
        "tags": [
          "EventoCuidado"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventoCuidado"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventoCuidado"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventoCuidado"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "EventoCuidado"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Veterinario/clinica/{clinicaId}": {
      "get": {
        "tags": [
          "Veterinario"
        ],
        "parameters": [
          {
            "name": "clinicaId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Veterinario"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Veterinario"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Veterinario"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnimalEstimacao": {
        "required": [
          "name",
          "species",
          "breed",
          "gender",
          "age",
          "weight",
          "birthDate",
          "tutoresId"
        ],
        "type": "object",
        "properties": {
          "animalEstimacaoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "species": {
            "type": "string"
          },
          "breed": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "age": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "weight": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "birthDate": {
            "type": "string",
            "format": "date-time"
          },
          "tutor": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Tutores"
              }
            ]
          },
          "tutoresId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "Clinica": {
        "required": [
          "tradeName",
          "cnpj",
          "email",
          "phoneNumber",
          "address"
        ],
        "type": "object",
        "properties": {
          "clinicaId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "tradeName": {
            "type": "string"
          },
          "cnpj": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "veterinarios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Veterinario"
            }
          }
        }
      },
      "EventoCuidado": {
        "required": [
          "animalEstimacaoId",
          "typeEvent",
          "description",
          "eventDate",
          "eventCompleted",
          "observations"
        ],
        "type": "object",
        "properties": {
          "eventoCuidadoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "animalEstimacaoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "animalEstimacao": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AnimalEstimacao"
              }
            ]
          },
          "typeEvent": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "eventDate": {
            "type": "string",
            "format": "date-time"
          },
          "eventCompleted": {
            "type": "boolean"
          },
          "observations": {
            "type": "string"
          }
        }
      },
...
      "Tutores": {
        "required": [
          "fullName",
          "cpf",
          "email",
          "password",
          "phoneNumber"
        ],
        "type": "object",
        "properties": {
          "tutoresId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "fullName": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "scoreEngagement": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "animaisEstimacao": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnimalEstimacao"
            }
          }
        }
      },
      "Veterinario": {
        "required": [
          "fullName",
          "crmv",
          "specialty",
          "email",
          "phoneNumber",
          "clinicaId"
        ],
        "type": "object",
        "properties": {
          "veterinarioId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "fullName": {
            "type": "string"
          },
          "crmv": {
            "type": "string"
          },
          "specialty": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "clinicaId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      }
    }
  },
...
}